requirements.txt
certifi==2020.12.5
chardet==4.0.0
ibm-cloud-sdk-core==3.9.0
ibm-vpc==0.5.1
idna==2.10
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
